+++ src/vector/ops/conversions/formats.rs
@@ -97,6 +97,21 @@ impl Geometry {
         Ok(wkt)
     }
 
+    /// Serialize the geometry as SFSQL 1.2 / ISO SQL / MM Part 3 WKT.
+    pub fn iso_wkt(&self) -> Result<String> {
+        let mut c_wkt = null_mut();
+        let rv = unsafe { gdal_sys::OGR_G_ExportToIsoWkt(self.c_geometry(), &mut c_wkt) };
+        if rv != OGRErr::OGRERR_NONE {
+            return Err(GdalError::OgrError {
+                err: rv,
+                method_name: "OGR_G_ExportToIsoWkt",
+            });
+        }
+        let wkt = _string(c_wkt).unwrap_or_default();
+        unsafe { gdal_sys::VSIFree(c_wkt as *mut c_void) };
+        Ok(wkt)
+    }
+
     /// Serializes the geometry to
     /// [WKB](https://en.wikipedia.org/wiki/Well-known_text_representation_of_geometry#Well-known_binary)
     /// (Well-Known Binary) format.
@@ -104,6 +119,8 @@ impl Geometry {
         let wkb_size = unsafe { gdal_sys::OGR_G_WkbSize(self.c_geometry()) as usize };
         // We default to little-endian for now. A WKB string explicitly indicates the byte
         // order, so this is not a problem for interoperability.
+        // 
+        // Consider using `Vec::MaybeUninit` in future.
         let byte_order = gdal_sys::OGRwkbByteOrder::wkbNDR;
         let mut wkb = vec![0; wkb_size];
         let rv =
@@ -112,6 +129,28 @@ impl Geometry {
             return Err(GdalError::OgrError {
                 err: rv,
                 method_name: "OGR_G_ExportToWkb",
+            });
+        }
+        Ok(wkb)
+    }
+
+    /// Serializes the geometry to SFSQL 1.2 / ISO SQL / MM Part 3
+    /// [WKB](https://en.wikipedia.org/wiki/Well-known_text_representation_of_geometry#Well-known_binary)
+    /// (Well-Known Binary) format.
+    pub fn iso_wkb(&self) -> Result<Vec<u8>> {
+        let wkb_size = unsafe { gdal_sys::OGR_G_WkbSize(self.c_geometry()) as usize };
+        // We default to little-endian for now. A WKB string explicitly indicates the byte
+        // order, so this is not a problem for interoperability.
+        //
+        // Consider using `Vec::MaybeUninit` in future.
+        let byte_order = gdal_sys::OGRwkbByteOrder::wkbNDR;
+        let mut wkb = vec![0; wkb_size];
+        let rv =
+            unsafe { gdal_sys::OGR_G_ExportToIsoWkb(self.c_geometry(), byte_order, wkb.as_mut_ptr()) };
+        if rv != gdal_sys::OGRErr::OGRERR_NONE {
+            return Err(GdalError::OgrError {
+                err: rv,
+                method_name: "OGR_G_ExportToIsoWkb",
             });
         }
         Ok(wkb)
@@ -137,6 +176,15 @@ mod tests {
         let wkt = "POLYGON ((45.0 45.0, 45.0 50.0, 50.0 50.0, 50.0 45.0, 45.0 45.0))";
         let orig_geom = Geometry::from_wkt(wkt).unwrap();
         let wkb = orig_geom.wkb().unwrap();
+        let new_geom = Geometry::from_wkb(&wkb).unwrap();
+        assert_eq!(new_geom, orig_geom);
+    }
+
+    #[test]
+    pub fn test_iso_wkb() {
+        let wkt = "POLYGON ZM ((45.0 45.0 0.0 0.0, 45.0 50.0 0.0 0.25, 50.0 50.0 0.0 0.50, 50.0 45.0 0.0 0.75, 45.0 45.0 0.0 1.0))";
+        let orig_geom = Geometry::from_wkt(wkt).unwrap();
+        let wkb = orig_geom.iso_wkb().unwrap();
         let new_geom = Geometry::from_wkb(&wkb).unwrap();
         assert_eq!(new_geom, orig_geom);
     }